pyproject.toml
[tool.poetry]
name = "api_keuangan"
version = "0.1.0"
description = ""
authors = ["bagussr <bagusssyamsu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
python-dotenv = "^0.20.0"
psycopg2 = "^2.9.3"
alembic = "^1.8.1"
SQLAlchemy = "^1.4.40"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
start = "api_keuangan.script:start"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
